Current statement selection implemented. Respects current selection, and if not able to select anything its selects all. Moved selection code to selection.ts.
